% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokeutils-package.R
\docType{package}
\name{smokeutils-package}
\alias{smokeutils}
\alias{smokeutils-package}
\title{smokeutils: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
